pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.4.12",
    "langchain-openrouter>=0.2.1",
    "flask>=3.1.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "langchain-openai>=1.1.12",
    "langfuse>=4.0.6",
]
